% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{add_nacolumn}
\alias{add_nacolumn}
\title{Filling missed columns with NAs}
\usage{
add_nacolumn(x, y)
}
\arguments{
\item{x}{(`data.frame`) The data frame to be compared.}

\item{y}{(`data.frame`) The data frame used as reference.}
}
\value{
A `data.frame`.
}
\description{
If columns of `y` are missed in `x`, the later gets these columns filled with
`NA` values.
}
\keyword{internal}
