% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{add_suffix}
\alias{add_suffix}
\title{Add a suffix when strings are identical}
\usage{
add_suffix(x, y, sep = "_")
}
\arguments{
\item{x}{(`character` of length 1) The name to be compared.}

\item{y}{(`character`) Existing names, with or without suffixes.}

\item{sep}{(`character` of length 1) Symbol to be placed between the original
name and the suffix.}
}
\value{
A `character` value, either `x` or `x` with suffix if already in `y`.
}
\description{
An integer will be added as suffix in `x` if an identical value is in `y`.
If a value with suffix is already in `y`, the function will search for the
highest suffix to avoid duplicated values.
}
\keyword{internal}
