% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{arrange_taxa}
\alias{arrange_taxa}
\title{Distribute parental taxa across taxon concepts}
\usage{
arrange_taxa(object)
}
\arguments{
\item{object}{(\code{taxlist}) A taxlist object including taxonomic ranks and parent-child relationships.}
}
\description{
A data frame including chains of parental taxa for each taxon concept.
}
\keyword{internal}
