% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_name.R
\name{collapse_names}
\alias{collapse_names}
\title{Collapse names}
\usage{
collapse_names(x, collapse)
}
\arguments{
\item{x}{A character vector containing a chain of names.}

\item{collapse}{A character value or vector used to collapse the names and
passed to \link{paste0}. If its lenght is 2, the second value will be used to
connect the two last names.}
}
\description{
Collapse name strings for document contents.
It will be used internally when formatting several names and including them
in a text body.
}
\keyword{internal}
