% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_taxa.R
\docType{methods}
\name{count_taxa}
\alias{count_taxa}
\alias{count_taxa,character-method}
\alias{count_taxa,factor-method}
\alias{count_taxa,taxlist-method}
\title{Count taxa within a taxlist object.}
\usage{
count_taxa(object, ...)

\S4method{count_taxa}{character}(object, rm.na = TRUE, ...)

\S4method{count_taxa}{factor}(object, rm.na = TRUE, ...)

\S4method{count_taxa}{taxlist}(object, level, ...)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{taxlist}}.}

\item{...}{further arguments passed among methods.}

\item{rm.na}{Logical value, whether NAs have to be removed from the input
vector or not.}

\item{level}{Character value indicating the taxonomic rank of counted taxa.}
}
\value{
An integer with the number of taxa.
}
\description{
Counting number of taxa within \code{\linkS4class{taxlist}} objects or
character vectors containing taxon names.
}
\details{
This function is written by convenience in order to reduce code for counting
taxa within \code{\linkS4class{taxlist}} objects and it is just a wrapper of
\code{\link{length}}.
}
\examples{
library(taxlist)

## factor method
count_taxa(iris$Species)

## taxlist method
count_taxa(Easplist)
count_taxa(Easplist, level="species")

}
\author{
Miguel Alvarez.
}
