% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matched_names-class.R
\name{matched_names-class}
\alias{matched_names-class}
\title{Names matched with a reference taxonomic list}
\description{
An S3 class containing results of names compared with a reference list.
This class enables further methods applied to these outputs, for instance
an interactive selection of multiple choices.
}
