\name{merge_species}
\docType{methods}
\alias{merge_species}
\alias{merge_species,taxlist-method}

\title{Merge Sub-Specific Taxa into Corresponding Species}
\usage{
\S4method{merge_species}{taxlist}(taxlist, ...)
}
\arguments{
\item{taxlist}{Object of class \code{taxlist}.}
\item{...}{Further arguments passed to or from other methods.}
}
\description{
This function aggregate all sub-specific syntaxa in the respective species.
}
\author{
Miguel Alvarez (\email{malvarez@uni-bonn.de}).
}
\examples{
## Add an example
}
