% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{overview_taxlist}
\alias{overview_taxlist}
\title{Function producing the overview of whole object.}
\usage{
overview_taxlist(object, units, check_validity)
}
\description{
Function producing the overview of whole object.
}
\keyword{internal}
