% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{overview_taxon}
\alias{overview_taxon}
\title{Function producing the overview of single taxon concepts.}
\usage{
overview_taxon(object, ConceptID, display, maxsum, secundum = NULL)
}
\description{
Function producing the overview of single taxon concepts.
}
\keyword{internal}
