% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{sort_backups}
\alias{sort_backups}
\title{Sorting files with time stamp and suffix}
\usage{
sort_backups(file, fext = ".rda")
}
\arguments{
\item{file}{A character value indicating the root of the backup's name. It
may include also the path relative to the working directory when the
backup is stored in a sub-folder.}

\item{fext}{A character value indicating the file extension (including the
dot symbol).}
}
\value{
A data frame including the sorted names of backup files from the
oldest to the newest.
}
\description{
When backups have been saved with a time stamp in the file's name and a
suffix, in the case of more than one backup in one day.
}
\keyword{internal}
