% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{tv.home}
\alias{tv.home}
\title{Guess Turboveg 2 installation path.}
\usage{
tv.home(recheck = FALSE)
}
\arguments{
\item{recheck}{(logical) reset even if option('tv_home') is already set.}
}
\value{
Reads and sets invisbly option('tv_home').
}
\description{
Reads and set invisibly \code{option("tv_home")}.
}
\author{
Florian Jansen
}
\keyword{internal}
