% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{two2one_df}
\alias{two2one_df}
\title{Inserting new rows and columns by merging two data frames}
\usage{
two2one_df(x, y)
}
\arguments{
\item{x, y}{(`data.frame`) The data frames to be merged.}
}
\description{
Two data frames sharing partially columns will be merged including the sum
of all variables.
}
\keyword{internal}
