% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{add_nacolumn}
\alias{add_nacolumn}
\title{Filling missed columns with NAs}
\usage{
add_nacolumn(x, y)
}
\arguments{
\item{x}{(\code{data.frame}) The data frame to be compared.}

\item{y}{(\code{data.frame}) The data frame used as reference.}
}
\value{
A \code{data.frame}.
}
\description{
If columns of \code{y} are missed in \code{x}, the later gets these columns filled with
\code{NA} values.
}
\keyword{internal}
