% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{add_suffix}
\alias{add_suffix}
\title{Add a suffix when strings are identical}
\usage{
add_suffix(x, y, sep = "_")
}
\arguments{
\item{x}{(\code{character} of length 1) The name to be compared.}

\item{y}{(\code{character}) Existing names, with or without suffixes.}

\item{sep}{(\code{character} of length 1) Symbol to be placed between the original
name and the suffix.}
}
\value{
A \code{character} value, either \code{x} or \code{x} with suffix if already in \code{y}.
}
\description{
An integer will be added as suffix in \code{x} if an identical value is in \code{y}.
If a value with suffix is already in \code{y}, the function will search for the
highest suffix to avoid duplicated values.
}
\keyword{internal}
