% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.list.R
\docType{methods}
\name{as.list}
\alias{as.list}
\alias{S4_to_list}
\alias{as.list,taxlist-method}
\title{Coerce an S4 object to a list.}
\usage{
S4_to_list(x)

\S4method{as.list}{taxlist}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\linkS4class{taxlist}} or any S4 class.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
An object of class \code{\link[base]{list}}.
}
\description{
Coercion of S4 objects to lists can be applied to explore their content,
avoiding errors caused by their validation.
}
\details{
The function `S4_to_list` transforms any S4 object to a list setting
slots to elements of the list and it is running internally in the method
`as.list` for \code{\linkS4class{taxlist}} objects.
}
\examples{
library(taxlist)
data(Easplist)
Easplist <- as.list(Easplist)
class(Easplist)

}
\author{
Miguel Alvarez.
}
