% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated-functions.R
\name{Deprecated-functions}
\alias{Deprecated-functions}
\alias{add_parent}
\alias{add_trait}
\alias{add_level}
\alias{replace_view}
\alias{taxlist2taxmap}
\alias{taxmap2taxlist}
\title{Deprecated functions}
\usage{
add_parent()

add_trait()

add_level()

replace_view()

taxlist2taxmap()

taxmap2taxlist()
}
\description{
Most of those functions have been replaced by alternative 'update' ones.
}
