\name{deprecated-functions}
\alias{add_parent}
\alias{add_trait}
\alias{add_level}

\title{Deprecated Functions}
\description{
For a more consistent structure of the package regarding first released
version, some functions were deprecated.
}
\usage{
add_parent()

add_trait()

add_level()
}
\seealso{
\code{\link{update_concept}}.
}
