% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Easplist-data.R
\docType{data}
\name{Easplist-data}
\alias{Easplist-data}
\alias{Easplist}
\title{List of vascular plants from East Africa}
\format{
An object of class \linkS4class{taxlist}.
}
\source{
\href{http://www.ville-ge.ch/musinfo/bd/cjb/africa/recherche.php}{African
Plant Database},
\href{http://www.givd.info/ID/AF-00-006}{SWEA-Dataveg}.
}
\usage{
Easplist
}
\description{
Example of an incomplete taxonomic list including taxa recorded in East
Africa.
}
\details{
This list is a subset of the taxonomic list implemented in the database
\href{http://www.givd.info/ID/AF-00-006}{SWEA-Dataveg}.
Since this list is being complemented regarding stored vegetation plots, it
is an incomplete list.
}
\examples{
summary(Easplist)

}
\keyword{datasets}
