% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels.R
\docType{methods}
\name{levels}
\alias{levels}
\alias{levels,taxlist-method}
\alias{levels<-}
\alias{levels<-,taxlist-method}
\title{Set and retrieves hierarchical levels}
\usage{
\S4method{levels}{taxlist}(x, ...)

\S4method{levels}{taxlist}(x) <- value
}
\arguments{
\item{x}{A \code{\linkS4class{taxlist}} object.}

\item{...}{Additional arguments passed among methods.}

\item{value}{A character vector with replacement values for levels o `x`.}
}
\value{
A `character` vector or a \code{\linkS4class{taxlist}} object with
added or modified taxonomic levels.
}
\description{
Taxonomic hierarchies can be set as levels in \code{\linkS4class{taxlist}}
objects, ordered from lower to higher levels.

Add taxonomic levels for specific taxon concepts in a
\code{\linkS4class{taxlist}} object.
Also changes in concept circumscription may implicate changes in its
taxonomic hierarchy.
}
\details{
Taxonomic levels will be handled as factors in the
\code{\linkS4class{taxlist}} objects.
Those levels are useful for creating subsets of related groups (e.g. by
functions \code{\link{get_children}} or \code{\link{get_parents}}).

Levels in combination to parent-child relationships will be further used for
checking consistency of taxonomic lists.

A replacement method of the form `levels(x) <- value` it is also implemented.
}
\examples{
## Get levels of species list
taxlist::levels(Easplist)

## Add aggregate as new taxonomic level
levels(Easplist) <- c("form","variety","subspecies","species",
    "complex", "aggregate","genus","family")
summary(Easplist)

}
\author{
Miguel Alvarez \email{kamapu78@gmail.com}
}
