% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxlist-package.R
\docType{package}
\name{taxlist-package}
\alias{taxlist}
\alias{taxlist-package}
\title{taxlist: Handling Taxonomic Lists}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Handling taxonomic lists through objects of class 'taxlist'. This package provides functions to import species lists from 'Turboveg' (\url{https://www.synbiosys.alterra.nl/turboveg/}) and the possibility to create backups from resulting R-objects. Also quick displays are implemented as summary-methods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cran.r-project.org/package=taxlist}
  \item \url{https://github.com/ropensci/taxlist}
  \item \url{https://docs.ropensci.org/taxlist/}
  \item Report bugs at \url{https://github.com/ropensci/taxlist/issues}
}

}
\author{
\strong{Maintainer}: Miguel Alvarez \email{kamapu78@gmail.com} (\href{https://orcid.org/0000-0003-1500-1834}{ORCID})

Other contributors:
\itemize{
  \item Zachary Foster \email{zacharyfoster1989@gmail.com} (\href{https://orcid.org/0000-0002-5075-0948}{ORCID}) [contributor]
  \item Sam Levin \email{levisc8@gmail.com} [reviewer]
  \item Margaret Siple \email{siplem@uw.edu} [reviewer]
}

}
\keyword{internal}
