% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxlist-package.R
\docType{package}
\name{taxlist-package}
\alias{taxlist-package}
\title{taxlist: Handling taxonomic lists.}
\description{
The class \linkS4class{taxlist} is defined in this package using the
S4 language.
The main task of \code{taxlist} objects is to keep the taxonomic coherence of
information included in taxonomic lists and to implement functions (methods)
for a proper data handling.
Objects of class \linkS4class{taxlist} can be included in further
objects, for instance in biodiversity records as done in the package
\href{https://cran.r-project.org/package=vegtable}{vegtable}.
}
\details{
The class \linkS4class{taxlist} is defined in this package using the
S4 language.
The main task of \code{taxlist} objects is to keep the taxonomic coherence of
information included in taxonomic lists and to implement functions (methods)
for a proper data handling.
Objects of class \linkS4class{taxlist} can be included in further
objects, for instance in biodiversity records as done in the package
\href{https://cran.r-project.org/package=vegtable}{vegtable}.

For a more detailed description of this package, see
\href{https://doi.org/10.1111/j.1654-1103.2010.01209.x}{Alvarez & Luebert
(2018)}.
}
\references{
\bold{Alvarez M, Luebert F (2018).} The taxlist package: managing plant
taxonomic lists in R. \emph{Biodiversity Data Journal} 6: e23635.
\url{https://doi.org/10.3897/bdj.6.e23635}
}
\author{
Miguel Alvarez \email{kamapu78@gmail.com}
}
